requirements.txt
flask>=3.0,<4.0
pyyaml>=6.0,<7.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
